refactor: GalleriesNearYou section
