+++ src/schema/v2/homeView/zones/legacy.ts
@@ -1,5 +1,6 @@
 import { ResolverContext } from "types/graphql"
-import { HomeViewSection, GalleriesNearYou } from "schema/v2/homeView/sections"
+import { HomeViewSection } from "schema/v2/homeView/sections"
+import { GalleriesNearYou } from "../sections/GalleriesNearYou"
 import { Auctions } from "../sections/Auctions"
 import { LatestArticles } from "../sections/LatestArticles"
 import { News } from "../sections/News"
